LayoutTests/fast/mediastream/SessionDescription-expected.txt

Tests the JSEP SessionDescription.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS sessionDescription = new SessionDescription(''); did not throw exception.
PASS sessionDescription.addCandidate() threw exception TypeError: Not enough arguments.
PASS sessionDescription.addCandidate(null) threw exception Error: TYPE_MISMATCH_ERR: DOM Exception 17.
PASS sessionDescription.addCandidate(undefined) threw exception Error: TYPE_MISMATCH_ERR: DOM Exception 17.
PASS sessionDescription.addCandidate(42) threw exception TypeError: Type error.
PASS iceCandidate = new IceCandidate('', ''); did not throw exception.
PASS sessionDescription.addCandidate(iceCandidate) did not throw exception.
PASS successfullyParsed is true

TEST COMPLETE
